♻️ Refactor: 파일 이름 변경 및 타입 분리
