+++ src/lib/core/hooks.type.ts
@@ -1,0 +1,9 @@
+import { VDOM, Component } from '../jsx/jsx-runtime.type';
+
+export interface Internals {
+  rootElement: HTMLElement | null;
+  rootComponent: Component | null;
+  currentVDOM: null | VDOM;
+  states: any[];
+  hookIndex: number;
+}
